rename reading challenge to goal
