+++ src/main/java/com/karankumar/bookproject/ui/goal/GoalView.java
@@ -1,4 +1,4 @@
-package com.karankumar.bookproject.ui.challenge;
+package com.karankumar.bookproject.ui.goal;
 
 import com.karankumar.bookproject.ui.MainView;
 import com.vaadin.flow.component.html.H3;
@@ -7,10 +7,10 @@ import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
 
-@Route(value = "readingChallenge", layout = MainView.class)
-@PageTitle("Reading Challenge | Book Project")
-public class ReadingChallengeView extends HorizontalLayout {
-    public ReadingChallengeView() {
+@Route(value = "goal", layout = MainView.class)
+@PageTitle("Goal | Book Project")
+public class GoalView extends HorizontalLayout {
+    public GoalView() {
         H3 text = new H3("Coming soon");
         VerticalLayout verticalLayout = new VerticalLayout(text);
         verticalLayout.setAlignItems(Alignment.CENTER);
